+++ features/support/preview_checker.rb
@@ -59,8 +59,8 @@ module PreviewChecker
   end
 
   def check_incentives
-    expect(page.body).to include('As a thank you, we will give your child')
-    expect(page).to have_tag('a.editable', text: 'a cash incentive of £10.50')
+    expect(page.body).to have_content('As a thank you, we will give your child')
+    expect(page).to have_tag('a.editable', text: 'a cash incentive of')
   end
 end