+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [FEATURE] - [Only have 1 button to 'toggle' if respondent's ability to consent](https://trello.com/c/Rk2I2yIg/266-2-only-have-1-button-to-toggle-if-respondents-ability-to-consent)
 * [FEATURE] - [Add guidance paragraph to each section of consent form builder](https://trello.com/c/Mrkkr3ij/306-2-add-guidance-paragraph-to-each-section-of-consent-form-builder)
 * [BUG] - [Word "child" on a form version for self-consent](https://trello.com/c/monsCFE9/309-1-bug-word-child-on-a-form-version-for-self-consent)
+* [FEATURE] - [Live preview: Incentives](https://trello.com/c/4sfX3wZj/263-3-live-preview-incentives)
 
 # 0.6.0 / 2017-12-14